Allow storing secrets in configuration files
Since we're using uWSGI and NGINX to serve requests in production,
figuring out what Django's actual environment is can be difficult. The
configuration of the environment is also not stored in an immediately
obvious place, making it easier to store all the secrets and
configuration variables in files. This change allows checking for a file
containing configuration information when the corresponding environment
variable is empty.